Provide alternative way of configuring jail.local
This provides a second, independent, way of configuring fail2ban. Instead of
copying in a valid template of the administrators making this builds a
configuration file out of variables supplied in configuration. The former way
allows migrating existing configuration files with relative ease while this
could be considered a better way as it benefits from ansibles variable
precedence.